rename base_metrics to shared_features, refactored for name change, made unit test file for detect_pen_lifts in shared_features
